# Conflicts:
#	src/main/java/poomasi/domain/auth/security/filter/JwtAuthenticationFilter.java
#	src/main/java/poomasi/domain/auth/security/handler/OAuth2SuccessHandler.java
#	src/main/java/poomasi/domain/auth/token/reissue/service/ReissueTokenService.java
#	src/main/java/poomasi/domain/farm/entity/Farm.java
#	src/main/java/poomasi/domain/reservation/dto/response/ReservationResponse.java
#	src/main/java/poomasi/domain/reservation/entity/Reservation.java
#	src/main/java/poomasi/domain/wishlist/service/WishListPlatformService.java
#	src/main/java/poomasi/payment/controller/PaymentController.java

+++ src/main/java/poomasi/domain/farm/dto/response/FarmResponse.java
@@ -25,7 +25,7 @@ public record FarmResponse(
                 .latitude(farm.getLatitude())
                 .longitude(farm.getLongitude())
                 .description(farm.getDescription())
-                .experiencePrice(farm.getExperiencePrice())
+                .experiencePrice(farm.getExperiencePrice().intValue())
                 .averageRating(farm.getAverageRating())
                 .build();
     }
